Consolidate spacing - Use CSS variables for all spacings - Remove Positions - Style hr globally
