+++ src/components/ContactInfo/ContactInfo.module.scss
@@ -1,0 +1,9 @@
+.contactInfo {
+  margin: var(--spacing--xs) 0;
+}
+
+.entry {
+  & + & {
+    margin-top: var(--spacing--m);
+  }
+}
