+++ src/components/ContactInfo/ContactInfoEntry/ContactInfoEntry.module.scss
@@ -2,7 +2,7 @@
 
 .contactInfoEntry {
   display: flex;
-  padding: 3px 0;
+  gap: var(--spacing--l);
 }
 
 .icon {
@@ -11,7 +11,6 @@
   flex: 0 0 auto;
   width: var(--size);
   height: var(--size);
-  margin-right: 12px;
   color: var(--color--icon);
 }
