+++ src/components/Experience/ExperienceEntry/Description/Description.module.scss
@@ -11,14 +11,15 @@
 
 .extra {
   flex: 0 0 auto;
-  margin-left: 40px;
+  margin-left: var(--spacing--xxxl);
+  text-align: right;
 
   &:before {
     content: '+ ';
   }
 
   @include mobile {
-    text-align: right;
     margin-left: 0;
+    margin-top: var(--spacing--s);
   }
 }
