+++ src/components/Experience/ExperienceEntry/PositionEntry/PositionEntry.module.scss
@@ -2,7 +2,7 @@
 
 .positionEntry {
   display: flex;
-  line-height: 1.6;
+  gap: var(--spacing--m);
 
   @include mobile {
     flex-direction: row-reverse;
@@ -11,7 +11,6 @@
 
 .timePeriod {
   flex: 0 0 88px;
-  margin-right: 10px;
 
   &:after {
     content: ')';
