--- src/components/Experience/ExperienceEntry/Positions/Positions.module.scss
@@ -1,3 +1,0 @@
-.positionEntry {
-  margin-bottom: 2px;
-}
