--- src/components/Experience/ExperienceEntry/Positions/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Positions';
