+++ src/pages/index.module.scss
@@ -5,6 +5,7 @@
 }
 
 .name {
+  margin-bottom: var(--spacing--xl);
   font-family: 'Lato', sans-serif;
   font-size: var(--font-size--xxxl);
   line-height: 1;
@@ -13,16 +14,14 @@
 }
 
 .description {
-  margin-top: 16px;
-  margin-bottom: 50px;
+  margin-bottom: var(--spacing--xxxl);
   font-size: var(--font-size--l);
   font-weight: 300;
   text-align: justify;
 
   @include mobile {
     font-size: var(--font-size--l);
-    line-height: 24px;
-    padding: 0 20px;
+    padding: 0 var(--spacing--xxl);
   }
 }
 
@@ -48,8 +47,7 @@
 }
 
 .sectionTitle {
-  margin-top: 0;
-  margin-bottom: 12px;
+  margin: var(--spacing--xl) 0;
   font-family: 'Lato', sans-serif;
   font-size: var(--font-size--xxl);
   font-weight: 300;
@@ -59,9 +57,9 @@
 .buttons {
   display: flex;
   justify-content: space-between;
-  gap: 20px;
+  gap: var(--spacing--xl);
   width: var(--a4--width);
-  margin-top: 40px;
+  margin-top: var(--spacing--xxxl);
 
   @media print {
     display: none;
@@ -79,7 +77,7 @@
 
 .buttonsRight {
   display: flex;
-  gap: 20px;
+  gap: var(--spacing--xxl);
 }
 
 .printButton {
