+++ src/styles/global.scss
@@ -47,3 +47,9 @@ h4 {
   -webkit-margin-after: 0;
   margin: 0;
 }
+
+hr {
+  margin: var(--spacing--m) 0;
+  border: none;
+  border-bottom: 1px dotted var(--color--separator);
+}
