+++ src/styles/variables.scss
@@ -8,7 +8,7 @@
   --spacing--m: 8px;
   --spacing--l: 12px;
   --spacing--xl: 16px;
-  --spacing--xxl: 20px;
+  --spacing--xxl: 24px;
   --spacing--xxxl: 40px;
 
   --font-size--s: 12px;
